+++ core/src/main/java/bisq/core/dao/node/full/rpc/dto/DtoPubKeyScript.java
@@ -17,9 +17,13 @@
 
 package bisq.core.dao.node.full.rpc.dto;
 
+import bisq.core.dao.state.model.blockchain.ScriptType;
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+import java.util.List;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -28,9 +32,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-@JsonPropertyOrder({"value", "n", "scriptPubKey"})
-public class RawOutput {
-    private Double value;
-    private Integer n;
-    private PubKeyScript scriptPubKey;
+@JsonPropertyOrder({"asm", "hex", "reqSigs", "type", "addresses"})
+public class DtoPubKeyScript {
+    private String asm;
+    private String hex;
+    private Integer reqSigs;
+    private ScriptType type;
+    private List<String> addresses;
 }
