+++ core/src/main/java/bisq/core/dao/node/full/rpc/dto/RawDtoOutput.java
@@ -17,13 +17,9 @@
 
 package bisq.core.dao.node.full.rpc.dto;
 
-import bisq.core.dao.state.model.blockchain.ScriptType;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-import java.util.List;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -32,11 +28,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-@JsonPropertyOrder({"asm", "hex", "reqSigs", "type", "addresses"})
-public class PubKeyScript {
-    private String asm;
-    private String hex;
-    private Integer reqSigs;
-    private ScriptType type;
-    private List<String> addresses;
+@JsonPropertyOrder({"value", "n", "scriptPubKey"})
+public class RawDtoOutput {
+    private Double value;
+    private Integer n;
+    private DtoPubKeyScript scriptPubKey;
 }
